pyproject.toml
[project]
name = "ais-uchet"
version = "2.0.0"
description = "АИС учета материальных средств"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "PyQt6>=6.6.0",
    "PyQt6-Qt6>=6.6.0",
    "PyQt6-sip>=13.6.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",  # Для хеширования паролей
    "openpyxl>=3.1.0",       # Работа с Excel
    "reportlab>=4.0.0",      # Генерация PDF
    "Pillow>=10.0.0",        # Работа с изображениями
    "python-dateutil>=2.8.2",
    "jsonschema>=4.19.0",    # Валидация JSON
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["E501"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
